Issue fixed for UK-26,28, 31, 38, 39, 42

+++ app/src/main/java/org/intelehealth/app/activities/pastMedicalHistoryActivity/PastMedicalHistoryActivity.java
@@ -605,13 +605,15 @@ public class PastMedicalHistoryActivity extends AppCompatActivity implements Que
                             try {
                                 // remove the visit
                                 VisitsDAO visitsDAO = new VisitsDAO();
-                                visitsDAO.deleteByVisitUUID(visitUuid);
+                                int count = visitsDAO.deleteByVisitUUID(visitUuid);
+                                if(count!=0) {
 
-                                ObsDAO obsDAO = new ObsDAO();
-                                obsDAO.deleteByEncounterUud(encounterAdultIntials);
-                                // remove the Encounter
-                                EncounterDAO encounterDAO = new EncounterDAO();
-                                encounterDAO.deleteByVisitUUID(visitUuid);
+                                    ObsDAO obsDAO = new ObsDAO();
+                                    obsDAO.deleteByEncounterUud(encounterAdultIntials);
+                                    // remove the Encounter
+                                    EncounterDAO encounterDAO = new EncounterDAO();
+                                    encounterDAO.deleteByVisitUUID(visitUuid);
+                                }
                                 setResult(RESULT_CANCELED);
                                 finish();
                             } catch (DAOException e) {
